Move "*Actions" interfaces into view package
