+++ richtextfx/src/main/java/org/fxmisc/richtext/ViewActions.java
@@ -8,9 +8,6 @@ import javafx.geometry.Point2D;
 import javafx.scene.Node;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.text.TextFlow;
-import org.fxmisc.richtext.model.Codec;
-import org.fxmisc.richtext.model.NavigationActions;
-import org.reactfx.util.Tuple2;
 import org.reactfx.value.Val;
 import org.reactfx.value.Var;
